Refactoring User Model Class and Wrapping the Password
Wrapping password to a new class to separate the responsability of the User class, now User class just have to Authenticate/Create a user.

Now Email and Passwords classes have to do their own validations.

By the last, i have refactored the Registration Controller to decrease the complexity and separate the responsabilitys of their methods.